data_mining/requirements.txt
pandas==2.1.4
numpy==1.24.4
psycopg2-binary==2.9.9
python-dotenv==1.0.1
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
jinja2==3.1.2
